Move configuration to the ShopBunny module
This way things work after reloading the app in development mode. Before this, when the Cart got reloaded on each request and therefore the Cart.shipping_cost_calculator was set to nil so things could not work.
By moving the configuration to a "save place" all is nice and dandy in development mode.