Really fixes the problem with differen to_json implementations
Uses a method now called "shop_bunny_json_attributes" on the Item classes
that can return whatever hash they want, which gets saved on CartItems as
JSON and CartItem#item is reconstruction an instance of the configured
Item class with this parsed JSON hash.

The default implementation uses just the ActiveRecord attributes method.